\name{nseries.MonteCarloSimulations}
\alias{nseriesInput.MonteCarloSimulations}
\alias{nseriesOutput.MonteCarloSimulations}

\title{Number of Series}
\description{Return the number of series.}
\usage{
    \method{nseriesInput}{MonteCarloSimulations}(x)
    \method{nseriesOutput}{MonteCarloSimulations}(x)
    }

\arguments{
    \item{x}{A featherForecasts object.}
    }
\value{An integer.}
\details{
    See the generic method.
    }
\keyword{programming}
\keyword{utilities}
\keyword{ts}

