\name{periodsMonteCarloSimulations}
\alias{periods.MonteCarloSimulations}
\alias{tframe.MonteCarloSimulations}

\title{Tframe or Number of Periods}
\description{Return the number of periods or the tframe.}
\usage{
    \method{periods}{MonteCarloSimulations}(x)
    \method{tframe}{MonteCarloSimulations}(x)
    }

\arguments{
    \item{x}{A MonteCarloSimulations object.}
    }
\value{An integer or a tframe object.}
\details{
    See the generic method.
    }
\keyword{programming}
\keyword{utilities}
\keyword{ts}

