\name{seriesNamesInput.MonteCarloSimulations}
\alias{seriesNamesInput.MonteCarloSimulations}
\alias{seriesNamesOutput.MonteCarloSimulations}

\title{TS Input and Output Specific Methods}
\description{See the generic function description.}
\usage{
    \method{seriesNamesInput}{MonteCarloSimulations}(x)
    \method{seriesNamesOutput}{MonteCarloSimulations}(x)
}
\arguments{
    \item{x}{an object from which to extract the names of the input or output
      series.}
}

\concept{DSE}
\keyword{ts}


