\name{TobsMonteCarloSimulations}
\alias{Tobs.MonteCarloSimulations}
\alias{tframe.MonteCarloSimulations}

\title{Tframe or Number of Observations}
\description{Return the number of Observations or the tframe.}
\usage{
    \method{Tobs}{MonteCarloSimulations}(x)
    \method{tframe}{MonteCarloSimulations}(x)
    }

\arguments{
    \item{x}{A MonteCarloSimulations object.}
    }
\value{An integer or a tframe object.}
\details{
    See the generic method.
    }
\keyword{programming}
\keyword{utilities}
\keyword{ts}

