\name{TSdata.coefEstEval}
\alias{TSestModel.coefEstEval}
\alias{TSmodel.coefEstEval}

\title{TS Extractor Specific Methods}
\description{See the generic function description.}
\usage{
    \method{TSestModel}{coefEstEval}(obj)
    \method{TSmodel}{coefEstEval}(obj, ...)
}
\arguments{
    \item{obj}{an object from which to extract the TSmodel or TSestModel.}
    \item{...}{arguments to be passed to other methods.}
}

\seealso{ 
    \code{\link[dse]{TSdata}} 
    \code{\link[dse]{TSestModel}} 
    \code{\link[dse]{TSmodel}} 
}

\concept{DSE}
\keyword{ts}

